// src/gql-client/__generated__/graphql.ts
export enum IssueState {
  /** An issue that has been closed */
  Closed = 'CLOSED',
  /** An issue that is still open */
  Open = 'OPEN'
}
// ...
export type GetRepositoryIssuesQuery = { __typename?: 'Query', repository?: { __typename?: 'Repository', issues: { __typename?: 'IssueConnection', totalCount: number, nodes?: Array<{ __typename?: 'Issue', state: IssueState, title: string, id: string, labels?: { __typename?: 'LabelConnection', nodes?: Array<{ __typename?: 'Label', color: string, name: string } | null> | null } | null, comments: { __typename?: 'IssueCommentConnection', totalCount: number } } | null> | null } } | null };